refactor(Archive): Move archive code out of archive button component
Move the code to the personal library component.
