+++ src/app/teacher/archive-projects-button/archive-projects-button.component.html
@@ -1,7 +1,7 @@
 <button
   mat-icon-button
   *ngIf="showArchive"
-  (click)="archiveProjects(true)"
+  (click)="archiveProjectsEvent.emit(true)"
   matTooltip="Archive"
   matTooltipPosition="above"
   i18n-matTooltip
@@ -11,7 +11,7 @@
 <button
   mat-icon-button
   *ngIf="!showArchive"
-  (click)="archiveProjects(false)"
+  (click)="archiveProjectsEvent.emit(false)"
   matTooltip="Restore"
   matTooltipPosition="above"
   i18n-matTooltip
