+++ src/app/teacher/teacher-run-list/teacher-run-list.component.spec.ts
@@ -72,7 +72,7 @@ class TeacherRunStub extends TeacherRun {
   }
 }
 
-describe('TeacherRunListComponent', () => {
+fdescribe('TeacherRunListComponent', () => {
   beforeEach(
     waitForAsync(() => {
       TestBed.configureTestingModule({
